fix: moved processBinding to CODE_PROCESSOR_PLUGIN
