+++ packages/core/src/generators/stencil/helpers/collect-class-string.ts
@@ -1,4 +1,3 @@
-import { ProcessBindingOptions } from '@/generators/stencil/helpers';
 import { MitosisNode } from '@/types/mitosis-node';
 
 const bindingOpenChar = '{';
@@ -6,11 +5,7 @@ const bindingCloseChar = '}';
 
 // This should really be a preprocessor mapping the `class` attribute binding based on what other values have
 // to make this more pluggable
-export const collectClassString = (
-  json: MitosisNode,
-  processBinding: (code: string, processBindingOptions: ProcessBindingOptions) => string,
-  processBindingOptions: ProcessBindingOptions,
-): string | null => {
+export const collectClassString = (json: MitosisNode): string | null => {
   const staticClasses: string[] = [];
 
   if (json.properties.class) {
@@ -24,11 +19,11 @@ export const collectClassString = (
 
   const dynamicClasses: string[] = [];
   if (typeof json.bindings.class?.code === 'string') {
-    dynamicClasses.push(processBinding(json.bindings.class.code, processBindingOptions));
+    dynamicClasses.push(json.bindings.class.code);
     delete json.bindings.class;
   }
   if (typeof json.bindings.className?.code === 'string') {
-    dynamicClasses.push(processBinding(json.bindings.className.code, processBindingOptions));
+    dynamicClasses.push(json.bindings.className.code);
     delete json.bindings.className;
   }
