+++ packages/core/src/generators/stencil/index.ts
@@ -1,1 +1,1 @@
-export * from './generate';
+export * from './component';
