+++ packages/core/src/generators/stencil/types.ts
@@ -13,4 +13,10 @@ export interface ToStencilOptions extends BaseTranspilerOptions {
   prefix?: string;
 }
 
-export interface StencilMetadata {}
+export interface StencilMetadata {
+  /**
+   * @deprecated Change the tag of your component.
+   * Note: Its deprecated change the class name of your component instead.
+   */
+  tagName?: string;
+}
